Combine ClusterServices table and Services table into single reusable component ServicesTable
