+++ assets/src/components/cd/services/Services.tsx
@@ -1,47 +1,15 @@
-import {
-  ComponentProps,
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react'
-import {
-  Chip,
-  EmptyState,
-  TabPanel,
-  Table,
-  useSetBreadcrumbs,
-} from '@pluralsh/design-system'
-import { useNavigate } from 'react-router'
+import { ComponentProps, useMemo, useState } from 'react'
+import { Chip, Table, useSetBreadcrumbs } from '@pluralsh/design-system'
 import { useTheme } from 'styled-components'
-import type { Row, TableState } from '@tanstack/react-table'
-import isEmpty from 'lodash/isEmpty'
-import { useDebounce } from '@react-hooks-library/core'
-import { type VirtualItem } from '@tanstack/react-virtual'
 
 import {
   AuthMethod,
   type ServiceDeploymentsRowFragment,
-  useServiceDeploymentsQuery,
 } from 'generated/graphql'
-import {
-  CD_REL_PATH,
-  SERVICES_REL_PATH,
-  getServiceDetailsPath,
-} from 'routes/cdRoutesConsts'
+import { CD_REL_PATH, SERVICES_REL_PATH } from 'routes/cdRoutesConsts'
 import { createMapperWithFallback } from 'utils/mapping'
-import { Edge, extendConnection } from 'utils/graphql'
-import { FullHeightTableWrap } from 'components/utils/layout/FullHeightTableWrap'
-import LoadingIndicator from 'components/utils/LoadingIndicator'
-import { GqlError } from 'components/utils/Alert'
-import { useSlicePolling } from 'components/utils/tableFetchHelpers'
 
-import {
-  CD_BASE_CRUMBS,
-  POLL_INTERVAL,
-  useSetCDHeaderContent,
-} from '../ContinuousDeployment'
+import { CD_BASE_CRUMBS, useSetCDHeaderContent } from '../ContinuousDeployment'
 
 import {
   ColActions,
@@ -54,7 +22,7 @@ import {
   ColStatus,
 } from './ServicesColumns'
 import { DeployService } from './deployModal/DeployService'
-import { ServicesFilters } from './ServicesFilters'
+import { ServicesTable } from './ServicesTable'
 
 export type ServicesCluster = Exclude<
   ServiceDeploymentsRowFragment['cluster'],
@@ -69,7 +37,7 @@ const authMethodToLabel = createMapperWithFallback<AuthMethod, string>(
   'Unknown'
 )
 
-const columns = [
+export const columns = [
   ColServiceDeployment,
   ColCluster,
   ColRepo,
@@ -131,164 +99,4 @@ export default function Services() {
   )
 
   return <ServicesTable setRefetch={setRefetch} />
-}
-
-function ServicesTable({
-  setRefetch,
-}: {
-  setRefetch?: (refetch: () => void) => void
-}) {
-  const theme = useTheme()
-  const navigate = useNavigate()
-  const [clusterId, setClusterId] = useState<string>('')
-  const [searchString, setSearchString] = useState()
-  const debouncedSearchString = useDebounce(searchString, 100)
-  const tabStateRef = useRef<any>(null)
-  const [virtualSlice, setVirtualSlice] = useState<
-    | {
-        start: VirtualItem | undefined
-        end: VirtualItem | undefined
-      }
-    | undefined
-  >()
-
-  const queryResult = useServiceDeploymentsQuery({
-    variables: {
-      ...(clusterId ? { clusterId } : {}),
-      q: debouncedSearchString,
-      first: SERVICES_QUERY_PAGE_SIZE,
-    },
-    fetchPolicy: 'cache-and-network',
-    // Important so loading will be updated on fetchMore to send to Table
-    notifyOnNetworkStatusChange: true,
-  })
-  const {
-    error,
-    fetchMore,
-    loading,
-    data: currentData,
-    previousData,
-  } = queryResult
-  const data = currentData || previousData
-  const serviceDeployments = data?.serviceDeployments
-  const pageInfo = serviceDeployments?.pageInfo
-  const { refetch } = useSlicePolling(queryResult, {
-    virtualSlice,
-    pageSize: SERVICES_QUERY_PAGE_SIZE,
-    key: 'serviceDeployments',
-    interval: POLL_INTERVAL,
-  })
-
-  useEffect(() => {
-    setRefetch?.(refetch)
-  }, [refetch, setRefetch])
-
-  const [tableFilters, setTableFilters] = useState<
-    Partial<Pick<TableState, 'globalFilter' | 'columnFilters'>>
-  >({
-    globalFilter: '',
-  })
-
-  const reactTableOptions: ComponentProps<typeof Table>['reactTableOptions'] =
-    useMemo(
-      () => ({
-        state: {
-          ...tableFilters,
-        },
-        meta: {
-          refetch,
-        },
-      }),
-      [refetch, tableFilters]
-    )
-
-  const fetchNextPage = useCallback(() => {
-    if (!pageInfo?.endCursor) {
-      return
-    }
-    fetchMore({
-      variables: { after: pageInfo.endCursor },
-      updateQuery: (prev, { fetchMoreResult }) =>
-        extendConnection(
-          prev,
-          fetchMoreResult.serviceDeployments,
-          'serviceDeployments'
-        ),
-    })
-  }, [fetchMore, pageInfo?.endCursor])
-
-  if (error) {
-    return <GqlError error={error} />
-  }
-  if (!data) {
-    return <LoadingIndicator />
-  }
-
-  const table = (
-    <div
-      css={{
-        display: 'flex',
-        flexDirection: 'column',
-        gap: theme.spacing.small,
-        height: '100%',
-      }}
-    >
-      <ServicesFilters
-        setTableFilters={setTableFilters}
-        searchString={searchString}
-        setSearchString={setSearchString}
-        clusterId={clusterId}
-        setClusterId={setClusterId}
-        tabStateRef={tabStateRef}
-        statusCounts={data.serviceStatuses}
-      />
-      <TabPanel
-        stateRef={tabStateRef}
-        css={{ height: '100%', overflow: 'hidden' }}
-      >
-        {!data ? (
-          <LoadingIndicator />
-        ) : !isEmpty(data?.serviceDeployments?.edges) ? (
-          <FullHeightTableWrap>
-            <Table
-              virtualizeRows
-              data={data?.serviceDeployments?.edges || []}
-              columns={columns}
-              css={{
-                maxHeight: 'unset',
-                height: '100%',
-              }}
-              onRowClick={(
-                _e,
-                { original }: Row<Edge<ServiceDeploymentsRowFragment>>
-              ) =>
-                navigate(
-                  getServiceDetailsPath({
-                    clusterId: original.node?.cluster?.id,
-                    serviceId: original.node?.id,
-                  })
-                )
-              }
-              hasNextPage={pageInfo?.hasNextPage}
-              fetchNextPage={fetchNextPage}
-              isFetchingNextPage={loading}
-              reactTableOptions={reactTableOptions}
-              reactVirtualOptions={SERVICES_REACT_VIRTUAL_OPTIONS}
-              onVirtualSliceChange={setVirtualSlice}
-            />
-          </FullHeightTableWrap>
-        ) : (
-          <div css={{ height: '100%' }}>
-            {searchString || clusterId ? (
-              <EmptyState message="No service deployments match your query." />
-            ) : (
-              <EmptyState message="Looks like you don't have any service deployments yet." />
-            )}
-          </div>
-        )}
-      </TabPanel>
-    </div>
-  )
-
-  return table
 }
